requirements.txt
streamlit
Pillow
openai
streamlit-geolocation